requirements.txt
# Core solver
ortools>=9.8

# Optional (commonly useful for I/O and config)
PyYAML>=6.0

# GUI
streamlit>=1.30
